Callbacks into OCaml run on the parent stack.
Previously, callbacks from C to OCaml allocated a fresh OCaml stack to
run the callback function. This isn't a problem usually, but it turns
out to be very bad for timer interrupts. Every time a timer interrupt is
handled, a new OCaml stack is created on the heap and discarded soon
after.

This commit modifies the behavior of callbacks such that they run on the
same stack as the parent, need only a check to see if there is enough
space in the parent stack to initiate a callback. Fixes #7.